+++ packages/use-store/src/types.ts
@@ -4,3 +4,6 @@ export type ReactStore<Value, Action = Value> = {
   [REACT_STORE_TYPE]: never;
   update: (action: Action) => void;
 };
+
+
+export type Reducer<S, A> = (state: S, action: A) => S;
